level_2/ORM/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
pydantic
faker
passlib
pyjwt
bcrypt